backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "numpy>=1.26.4",
    "openai==1.3.5",
    "pillow>=10.2.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.0",
    "supabase==2.0.3",
    "torch>=2.0.0",
    "uvicorn[standard]==0.24.0",
]
